+++ src/scripts/setup/options.ts
@@ -14,5 +14,6 @@ export const configTypeOptions: Option<ConfigType>[] = [
     { label: 'Prettier', value: 'prettier' },
     { label: 'ESLint', value: 'eslint' },
     { label: 'TypeScript', value: 'typescript' },
-    { label: 'Cursor Settings/Rules', value: 'editor' },
+    { label: 'Cursor Settings', value: 'cursorSettings' },
+    { label: 'Agent Rules & Skills', value: 'agentRulesAndSkills' },
 ];