+++ src/types/index.ts
@@ -1,5 +1,10 @@
 export type ProjectType = 'node' | 'reactNative' | 'reactNext' | 'reactTanStackStart' | 'reactVite';
-export type ConfigType = 'eslint' | 'prettier' | 'typescript' | 'editor';
+export type ConfigType =
+    | 'eslint'
+    | 'prettier'
+    | 'typescript'
+    | 'cursorSettings'
+    | 'agentRulesAndSkills';
 
 export type PackageManager = 'bun' | 'npm' | 'pnpm' | 'yarn';